requirements.txt
euclid
pycairo
